# Conflicts:
#	i21Database/Scripts/Post-Deployment/SM/DefaultData/4_StartingNumbers.sql
#	i21Database/i21Database.sqlproj

+++ i21Database/dbo/Views/vyuTMDeliveryFillReport.sql
@@ -94,7 +94,7 @@ SELECT
 	, O.strRouteId
 	, P.strFillMethod
 	, strBetweenDlvry = (CASE WHEN C.intFillMethodId = U.intFillMethodId THEN CONVERT(VARCHAR,C.dtmNextDeliveryDate,101)
-							ELSE CAST((CONVERT(INT,C.dblDegreeDayBetweenDelivery)) AS NVARCHAR(10))
+							ELSE CAST((CONVERT(NUMERIC(18,2),C.dblDegreeDayBetweenDelivery)) AS NVARCHAR(10))
 						END)  
 	, strLocation = (CASE WHEN ISNUMERIC(C.strLocation) = 1 THEN C.strLocation
 					 ELSE SUBSTRING(C.strLocation, PATINDEX('%[^0]%',C.strLocation), 50)
